\name{HamiltonGDP}
\docType{data}
\alias{HamiltonGDP}
\title{Quarterly U.S. GDP Growth, 1952Q3-1984Q4}

\description{Hamilton's (1989) quarterly data on U.S. GDP growth.
}

\usage{data(HamiltonGDP)}

\format{\code{ts} object of quarterly GDP growth named \code{gdp}. This
  \code{ts} classed object has 130 observations. 
}

\source{
  Hamilton, James. 1989. "A new approach to the economic analysis of
  nonstationary time series and the business cycle." Econmetrica,
  357--384. 
}

\references{
  Hamilton, James. 1989. "A new approach to the economic analysis of
  nonstationary time series and the business cycle." Econmetrica,
  357--384. 
}

\keyword{datasets}
