% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.msvar.R
\name{residuals.msvar}
\alias{residuals.msvar}
\title{Calculate weighted residuals over the different regimes based on their probability}
\usage{
\method{residuals}{msvar}(x)
}
\description{
Calculate weighted residuals over the different regimes based on their probability
}
